requirements.txt
Django==4.1.3
django_environ==0.9.0
django_taggit==3.1.0
environ==1.0
